fix(#2502): updated sidebar labels for 1.5.0
